fix: PostAdd/EditDialog 이름 변경
