+++ src/widgets/post/index.ts
@@ -1,5 +1,5 @@
 export { default as PostTable } from "./PostTable";
 export { default as PostTableRow } from "./PostTableRow";
-export { default as PostAddDialog } from "./PostAddDialog";
 export { default as PostEditDialog } from "./PostEditDialog";
+export { default as PostAddDialog } from "./PostAddDialog";
 export { default as PostDetailDialog } from "./PostDetailDialog";
